경로가 나오긴 한다. 이대로 움직여 보자

+++ 숙제/3개월차/API 시연회/RimWorld/RimWorld/TileMgr.cpp
@@ -77,7 +77,7 @@ void CTileMgr::Release()
 
 void CTileMgr::Set_TileOption(float _fX, float _fY, TILEOPT _eOpt)
 {
-	int _iIdx = _fX / TILECX + _fY / TILECY * TILEX;
+	int _iIdx = int(_fX / TILECX) + int(_fY / TILECY * TILEX);
 	static_cast<CTile*>(m_arrTile.at(_iIdx))->Set_Option(_eOpt);
 }
